+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [7.2.0] (IN PROGRESS)
 
+* Add tests for useFetchExportTitlesFromPackage hook. (UIEH-1182)
+
 ## [7.1.1] (https://github.com/folio-org/ui-eholdings/tree/v7.1.0) (2022-03-21)
 
 * Settings > eholdings > Edit knowledge base credentials > Updating any field value generates an error message BUT the saves the update. (UIEH-1247)